Add GetLastTemplate,Remove CoinbaseOutputDataSize, Remove SetNonce, Add SetCoinbaseOutput
1. In order to send TxShortHashList to the JDC the TP MUST know the coinbase
2. The nonce used to build the tx_short_hash_list should be changed for every NewTemplate

Solving 2 is easy we just remove the SetNonce message and we make the TP chose the nonce and send it to the JDC in the  TxShortHashList message

Solving 1 is harder: the TP have CoinbaseOutputDataSize message so that pool can set up a maximum size and then change the coinbase without the need to update the TP. This is useful cause it avoid the creation of invalid NewTemplate and reduce latency. But when we do job declaration the TP needs to know the coinbase in order to create the TxShortHashList  that MUST be sent along the  NewTemplate . What we propose to do is to change the CoinbaseOutputDataSize  into SetCoinbaseOutput  and to add a new field in NewTemplate called coinbase (that is needed by the client in oder to check which coinbase has been used by the TP to create the template). We also propose to add a new message to the TDP called GetLastTemplate, when the server receive it, it must  create a NewTemplate using the last received coinbase.