replace async terminology with coinbase vs template modes..
as agreed between @GitGab19, @TheBlueMatt, @Fi3 and @plebhash, we are
dropping the "async" terminology in preference of coinbase vs template
"modes".

async terminology was causing confusion and leading to conflicting interpretation of the specs.

now, coinbase vs template modes make it explicitly clear how each side (JDC vs JDS) handles job declaration, especially with regards to:
- knowledge about the txdata contained in the template
- ability to propagate a valid block

the original semantics of async is now expressed with the notion of "optimistic mining"