re-introduce min_ntime to SetCustomMiningJob
SetCustomMiningJob cannot ever be future job.

on 0fd5cf5d50ed0aa10339c3a676496acf536af94c we completely dropped this field, which is not correct.

while SetCustomMiningJob.min_ntime field cannot be optional, it cannot be dropped... it needs to be mandatory

otherwise pool could reward work for custom jobs on timestamps that are in the past (which would lead to rejected blocks)