Disallow declaring future templates
Add prev_hash to DeclareMiningJob.

Support for future templates can result in the JDS failing to reconstruct a block around the time of a tip update. Since it doesn't know the prev_hash, it guess its own tip. When there is a height mismatch, this results in a BIP34 validation failure.

Fix this by requiring the JDC to hold on to future templates until the prevhash is known.

Also add a few error strings for DeclareMiningJob.Error.